split scenarios and tests
